实验的要求都over了 当然 有各种可以提升的地方 统计那里也可以加好多 比如排名 关键字排序之类的

+++ 从零开始的win32sdk开发/filter.cpp
@@ -20,22 +20,22 @@ set<int> filter::return_del_item(string keystring,int flitem,string flstring)
 bool filter::cmp0(string x,string cmp)
 {
 	if(ifstring(x))return x<cmp;
-	else return atoi(x.c_str())<atoi(cmp.c_str());
+	else return atof(x.c_str())<atof(cmp.c_str());
 }
 bool filter::cmp1(string x,string cmp)
 {
 	if(ifstring(x))return x<=cmp;
-	else return atoi(x.c_str())<=atoi(cmp.c_str());
+	else return atof(x.c_str())<=atof(cmp.c_str());
 }
 bool filter::cmp2(string x,string cmp)
 {
 	if(ifstring(x))return x>cmp;
-	else return atoi(x.c_str())>atoi(cmp.c_str());
+	else return atof(x.c_str())>atof(cmp.c_str());
 }
 bool filter::cmp3(string x,string cmp)
 {
 	if(ifstring(x))return x>=cmp;
-	else return atoi(x.c_str())>=atoi(cmp.c_str());
+	else return atof(x.c_str())>=atof(cmp.c_str());
 }
 bool filter::cmp4(string x,string cmp)
 {
@@ -49,7 +49,7 @@ bool filter::cmp5(string x,string cmp)
 }
 bool filter::ifstring(string x)
 {
-	int xc=atoi(x.c_str());
+	double xc=atof(x.c_str());
 	if(xc!=0||to_string(xc)==x)return false;
 	else return true;
 }
